move dir and change browserRouter to hashRouter
