+++ src/assets/style/CSBlogPage.css
@@ -180,6 +180,8 @@
     align-items: flex-start;
     box-shadow: 0px 4px 12px 9px rgba(247, 185, 59, 0.25);
     border-radius: 17px;    
+    margin: 0 2rem;
+    margin-bottom: 2rem;
     height: 400px;
     max-width: 365px;
 }
